forkServer error handler for socket accept
It's not necessary that the server should stop if accepting a connection
throws an exception.

Currently, it still does, as we use `throwIO` for this error handler,
which causes the transport's termination handler to run and kill all
service. We should think about a better policy here, perhaps wait 1
second and try again. An example case: the process has no more file
descriptors available, so `accept` will always throw an exception; the
server could keep on trying intermittently until some file descriptors
are closed.